requirements.txt
Flask==3.1.3
requests==2.32.5
python-dotenv==1.2.1
Pillow==12.1.1
cachetools==5.5.2
redis==5.2.1
gunicorn==23.0.0
